% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{apply.sum.effects}
\alias{apply.sum.effects}
\title{Apply the sum of effects to each participant.}
\usage{
apply.sum.effects(data.dt, sum.effects.dt)
}
\arguments{
\item{data.dt}{data.table with minimally participant ID and outcome.}

\item{sum.effects.dt}{The data.table generated by calculate.sum.effects.dt,
including an effect to be applied to each participant.}
}
\value{
A data.table including participant IDs, and outcomes after summed
  effects are applied.
}
\description{
Apply the sum of effects to each participant.
}
