% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{calculate.icc.dt.of.simulations}
\alias{calculate.icc.dt.of.simulations}
\title{Generates a number of simulated datasets, and assesses the ICC of each.
Parameter lists are as provided to generate.sim.parameters.dt and
generate.confound.interventions.parameters.dt, but if there are vectors for
any of the values, these will be randomly sampled, generating multiple
scenarios.}
\usage{
calculate.icc.dt.of.simulations(
  cluster.dt,
  sequence.dt,
  sim.parameters.args.list,
  confound.interventions.args.list,
  n.iterations = 20,
  progress.bar = T,
  binarise.data = F,
  max.time = NULL
)
}
\arguments{
\item{cluster.dt}{data.table with the correspondence between cluster and
sequence, with columns cluster and sequence.}

\item{sequence.dt}{data.table with information about the sequences, with
columns sequence, transition.time, and intervention.time.}

\item{sim.parameters.args.list}{List of parameters, as provided to
generate.sim.parameters.dt}

\item{confound.interventions.args.list}{List of parameters, as provided to
confound.interventions.args.list}

\item{n.iterations}{Integer number of *total* iterations to assess (i.e. not
the number per scenario)}

\item{binary.data}{Generate binary data (Default: F)}
}
\value{
A data.table with a column for each simulation parameter, and a
  column named ICC added.
}
\description{
Generates a number of simulated datasets, and assesses the ICC of each.
Parameter lists are as provided to generate.sim.parameters.dt and
generate.confound.interventions.parameters.dt, but if there are vectors for
any of the values, these will be randomly sampled, generating multiple
scenarios.
}
