% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{calculate.icc}
\alias{calculate.icc}
\title{Calculates intracluster correlation coefficient for observations}
\usage{
calculate.icc(
  data.dt,
  mode = c("binary", "continuous", "auto")[3],
  return.type = c("value", "variables")[1]
)
}
\arguments{
\item{data.dt}{data.table with minimally columns for cluster and outcome.}

\item{mode}{Type of variable for which ICC will be calculated. Can be
automatically detected on the basis of number of levels. (Default: auto)}

\item{return.type}{Can return either the single value rho, or a list of all
values calculated (Default: 'value')}
}
\value{
Numeric ICC
}
\description{
Calculates intracluster correlation coefficient for observations
}
