% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{calculate.p.from.stat.dt}
\alias{calculate.p.from.stat.dt}
\title{Calculate a p-value from a list of statistics, try and automatically detect
the appropriate one. If there is a column named 'wgt', use that to weight the
average per permutation number.}
\usage{
calculate.p.from.stat.dt(stat.dt)
}
\arguments{
\item{stat.dt}{data.table with a column called perm.num which has the
comparison from the actual data with 0}

\item{stat.col}{character name of the column which is the stat}
}
\value{
Two-sided p-value
}
\description{
Calculate a p-value from a list of statistics, try and automatically detect
the appropriate one. If there is a column named 'wgt', use that to weight the
average per permutation number.
}
