% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{calculate.predictor.value.from.regression.model}
\alias{calculate.predictor.value.from.regression.model}
\title{Takes a linear regression model and provided values of predictors and
outcome, and calculates a provided (or missing) predictor (or outcome).
Probably doesn't work for models with interaction terms.}
\usage{
calculate.predictor.value.from.regression.model(
  model,
  constant.values.list,
  predict.var = NULL
)
}
\arguments{
\item{model}{A lm generated model.}

\item{constant.values.list}{List of provided values for outcome and
predictors.}

\item{predict.var}{Character name of the variable to predict. If NULL the
function will look for a missing item in outcome.values in
constant.values.list. If predict.var exists in constant.values.list, its
value there will be ignored (Default : NULL)}
}
\value{
A scatter ggplot
}
\description{
Takes a linear regression model and provided values of predictors and
outcome, and calculates a provided (or missing) predictor (or outcome).
Probably doesn't work for models with interaction terms.
}
