% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{calculate.sum.effects.dt}
\alias{calculate.sum.effects.dt}
\title{Sums the effects for each participant.}
\usage{
calculate.sum.effects.dt(effects.dt)
}
\arguments{
\item{effects.dt}{The data.table generated by generate.effects.dt}
}
\value{
A data.table including participant IDs, and summed effects.
}
\description{
Sums the effects for each participant.
}
