% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{cols.exceed.level.threshold}
\alias{cols.exceed.level.threshold}
\title{Detects what columns in a data.table have more than a certain number of
columns.}
\usage{
cols.exceed.level.threshold(
  data.dt,
  threshold = 1,
  cols.to.ignore = c("iteration", "icc")
)
}
\arguments{
\item{data.dt}{data.table with minimally participant ID and outcome.}

\item{threshold}{The number of levels over which the column name will be
returned (Default: 1)}

\item{cols.to.ignore}{Columns to not count (Default: c('iteration', 'icc'))}
}
\value{
A vector of strings which are the names of columns with numers of
  levels exceeding the threshold.
}
\description{
Detects what columns in a data.table have more than a certain number of
columns.
}
