% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{convert.continuous.to.binary}
\alias{convert.continuous.to.binary}
\title{Converts a continuous variable to a binary one. Zero or negative values of
the continuous variable will be considered to have the first level of the
variable.}
\usage{
convert.continuous.to.binary(
  data.dt,
  max.chance = 0.5,
  var.levels = c(0, 1),
  delete.outcome.orig = T,
  delete.prob = T
)
}
\arguments{
\item{data.dt}{data.table with minimally participant ID and outcome.}

\item{max.chance}{The chance that the highest value of the continuous
variable will be assigned the second level of the binary variable.
(Default: 0.5)}

\item{var.levels}{Vector containing the values that the binary outcome will
assume (Default: c(0,1))}

\item{delete.outcome.orig}{Delete original continuous outcome (Default: TRUE)}

\item{delete.prob}{Delete calculated probability of var.levels[2] (Default:
TRUE)}
}
\value{
A data.table with the continuous outcome column replaced by the
  specified binary variable. Not necessary to use return value, as data.dt
  will be modified by reference,
}
\description{
Converts a continuous variable to a binary one. Zero or negative values of
the continuous variable will be considered to have the first level of the
variable.
}
