% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{cut.data.dt.into.conditions}
\alias{cut.data.dt.into.conditions}
\title{Assigns control, transition, and intervention condition column on the basis
of experiment info. Can optionally include transition period data points.}
\usage{
\method{cut}{data.dt.into.conditions}(
  data.dt,
  cluster.dt,
  sequence.dt,
  exclude.transition = T,
  cut.labels = c("control", "transition", "intervention")
)
}
\arguments{
\item{data.dt}{data.table with columns participant, cluster, time, and
outcome. Outcome should be continuous.}

\item{cluster.dt}{data.table with the correspondence between cluster and
sequence, with columns cluster and sequence.}

\item{sequence.dt}{data.table with information about the sequences, with
columns sequence, transition.time, and intervention.time.}

\item{exclude.transition}{boolean, should the result exclude data points from
the transition period? (Default = T)}

\item{cut.labels}{Character vector of labels for control, transition, and
intervention periods. Probably don't change this (Default:
c('control','transition','intervention'))}
}
\value{
The input data.dt with conditions attached.
}
\description{
Assigns control, transition, and intervention condition column on the basis
of experiment info. Can optionally include transition period data points.
}
