% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{cut.data.dt.into.periods}
\alias{cut.data.dt.into.periods}
\title{Make a period column in a data.table, with each period starting at an
intervention time.}
\usage{
\method{cut}{data.dt.into.periods}(data.dt, sequence.dt)
}
\arguments{
\item{data.dt}{data.table with columns participant, cluster, time, and
outcome. Outcome should be continuous.}

\item{sequence.dt}{data.table with information about the sequences, with
columns sequence, transition.time, and intervention.time.}
}
\value{
A data.table with period column assigned. This will be done by
  reference for speed, so you don't need to use the return value.
}
\description{
Make a period column in a data.table, with each period starting at an
intervention time.
}
