% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{cut.time.into.conditions}
\alias{cut.time.into.conditions}
\title{Cut timepoints from a single cluster into separate conditions, return list of
conditions.}
\usage{
\method{cut}{time.into.conditions}(
  time,
  transition.time,
  intervention.time,
  time.start = -Inf,
  time.end = Inf,
  cut.labels = c("control", "transition", "intervention")
)
}
\arguments{
\item{time}{numeric vector of times.}

\item{transition.time}{numeric indicating when the transition period should
start.}

\item{intervention.time}{numeric indicating when the intervention period
should start.}

\item{time.start}{numeric indicating when labelling the transition time
should start. (Default = -Inf)}

\item{time.end}{numeric indicating when labelling the intervention time
should end. (Default = Inf)}

\item{cut.labels}{What should the labels be? (Default:
c('control','transition','intervention'))}
}
\value{
List of conditions in the same order as presented.
}
\description{
Cut timepoints from a single cluster into separate conditions, return list of
conditions.
}
