% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.helpers.R
\name{elapsed.time.message}
\alias{elapsed.time.message}
\title{Calculates and outputs a message of elapsed time.}
\usage{
elapsed.time.message(start.time, n.iterations = NULL, progress.bar = NULL)
}
\arguments{
\item{start.time}{Numeric start time from Sys.time()}

\item{n.iterations}{Integer how many iterations (time per iteration won't be
calculated if NULL, Default: NULL)}

\item{progress.bar}{Progress bar to close (ignored if NULL, Default: NULL)}
}
\description{
Calculates and outputs a message of elapsed time.
}
