% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{faceted.line.plot}
\alias{faceted.line.plot}
\title{Plots a line and scatter plot faceted by sequence and cluster with a few
visual aids. If it's a permutation, maybe provide the actual clusters so that
the theming can work properly. sort the input by outcome to make it faster.}
\usage{
faceted.line.plot(
  data.dt,
  sequence.dt,
  cluster.dt,
  actual.cluster.dt = NULL,
  min.time = 0,
  max.time = NULL
)
}
\arguments{
\item{data.dt}{data.table with columns participant, cluster, time, and
outcome. Outcome should be continuous.}

\item{sequence.dt}{data.table with information about the sequences, with
columns sequence, transition.time, and intervention.time.}

\item{cluster.dt}{data.table with the correspondence between cluster and
sequence, with columns cluster and sequence.}

\item{actual.cluster.dt}{This function is partly designed to visualise
permutation statistics, if cluster.dt is permuted, then you might provide a
similar data.table with the actual correpsondence between cluster and
sequence so that the graph can be themed as such (i.e. data colours
matching original sequence).}

\item{min.time}{Numeric minimum time on x axis (Default: 0)}

\item{max.time}{Numeric maximum time on x axis, will try to automatically
generate from sequence.dt if NULL (Default: NULL)}
}
\value{
A ggplot with data faceted by sequence and cluster.
}
\description{
Plots a line and scatter plot faceted by sequence and cluster with a few
visual aids. If it's a permutation, maybe provide the actual clusters so that
the theming can work properly. sort the input by outcome to make it faster.
}
