% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{fit.icc.model}
\alias{fit.icc.model}
\title{Creates a multiple regression model predicting ICC so that data can be
simulated fulfilling criteria.}
\usage{
fit.icc.model(
  target.icc = 0.02,
  sim.parameters.args.list = NULL,
  confound.interventions.args.list = NULL,
  interactions.list = NULL
)
}
\arguments{
\item{target.icc}{Numeric target ICC ([0-1], Default: 0.02)}

\item{sim.parameters.args.list}{A list of the arguments for
generate.sim.parameters.dt, with names being the argument names, and values
being the values, with vectors for the values that should be optimised.}

\item{confound.interventions.args.list}{A list of the arguments for
generate.confound.interventions.parameters.dt, with names being the argument names,
and values being the values, with vectors for the values that should be
optimised.}

\item{interactions.list}{A list of lists of interactions to assess (Default:
none)}

\item{iterations}{Number of ICC that will be calculated to fit the regression
model (Default: 500)}
}
\value{
A regression model with the provided predictors and ICC as the
  outcome.
}
\description{
Creates a multiple regression model predicting ICC so that data can be
simulated fulfilling criteria.
}
