% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{fit.regression.model}
\alias{fit.regression.model}
\title{Fits a regression model predicting an outcome variable from provided variables.}
\usage{
fit.regression.model(
  icc.dt,
  predictor.vars = NULL,
  outcome.var = "icc",
  include.interactions = F,
  model.type = c("lm", "loess")[1]
)
}
\arguments{
\item{icc.dt}{A data.table with the parameters of the various simulations and
ICC.}

\item{predictor.vars}{Calculate interation effects, or just simple effects
(Default: F)}

\item{outcome.var}{Character column name to predict (Default: ICC)}

\item{model.type}{Type of model to fit, currently supports linear ('lm') and
quadratic ('loess') (Default: 'lm')}
}
\value{
A linear model object.
}
\description{
Fits a regression model predicting an outcome variable from provided variables.
}
