% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{generate.base.data.dt}
\alias{generate.base.data.dt}
\title{Generate a data table populated with participants from simulated recruitment
rates,}
\usage{
generate.base.data.dt(sim.parameters.dt, min.time = 0, max.time = NULL)
}
\arguments{
\item{sim.parameters.dt}{The data.table generated by
generate.sim.parameters.dt}
}
\value{
A data.table including participant IDs, cluster, a base outcome and
  time.
}
\description{
Generate a data table populated with participants from simulated recruitment
rates,
}
