% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{generate.cluster.dt}
\alias{generate.cluster.dt}
\title{Generates a data.table that describes how clusters are assigned to sequences.}
\usage{
generate.cluster.dt(sequence.dt, n.cluster.per.sequence = 1)
}
\arguments{
\item{sequence.dt}{data.table with information about the sequences, with
columns sequence, transition.time, and intervention.time.}

\item{n.cluster.per.sequence}{Integer number of clusters assigned to each
sequence.}
}
\value{
A data.table with cluster and sequence ID columns.
}
\description{
Generates a data.table that describes how clusters are assigned to sequences.
}
