% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{generate.confound.interventions.parameters.dt}
\alias{generate.confound.interventions.parameters.dt}
\title{Generates a table of confounding interventions (e.g. those from other
interventions or such). There can be a potentially random number of
confounding interventions.}
\usage{
generate.confound.interventions.parameters.dt(
  cluster.dt,
  min.time = 0,
  max.time = 0,
  n.confound.intervention.mean = 5,
  n.confound.intervention.sd = 1,
  confound.intervention.transition.duration.mean = 5,
  confound.intervention.transition.duration.sd = 0.6,
  confound.intervention.mean = 0.7,
  confound.intervention.sd = 0.1,
  n.to.generate = 1
)
}
\arguments{
\item{cluster.dt}{Earliest time of a random effect (Default: 0)}

\item{min.time}{Earliest time of a random effect (Default: 0)}

\item{max.time}{Latest time of a random effect (Default: 0)}

\item{n.confound.intervention.mean}{How many random effects total, mean parameter for
rnorm (Default: 5)}

\item{n.confound.intervention.sd}{How many random effects total, sd parameter for
rnorm (Default: 1)}

\item{confound.intervention.transition.duration.mean}{Variance of time taken for
transition of effect between random effects (Default: 5)}

\item{confound.intervention.mean}{Mean random effect (Default: 0.7)}

\item{confound.intervention.sd}{Variance of the random effects (Default: 0.1)}

\item{n.to.generate}{Integer how many different effects to generate. If more
than one, an integer column will be added (sim.number) to delineate.
(Default: 1)}
}
\value{
A numeric vector containing the effect at each provided time point.
}
\description{
Generates a table of confounding interventions (e.g. those from other
interventions or such). There can be a potentially random number of
confounding interventions.
}
