% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{generate.data.dt}
\alias{generate.data.dt}
\title{Creates a data.table with participants and simulates data for them.}
\usage{
generate.data.dt(
  cluster.dt,
  sequence.dt,
  sim.parameters.dt,
  confound.interventions.parameters.dt,
  return.intermediaries = F,
  max.time = NULL
)
}
\arguments{
\item{cluster.dt}{data.table with the correspondence between cluster and
sequence, with columns cluster and sequence.}

\item{sequence.dt}{data.table with information about the sequences, with
columns sequence, transition.time, and intervention.time.}

\item{sim.parameters.dt}{The data.table generated by
generate.sim.parameters.dt}

\item{confound.interventions.parameters.dt}{The data.table generated by
generate.confound.interventions.parameters.dt}

\item{return.intermediaries}{If true, intermediaries will be returned (return
type will be list), useful for plotting.}
}
\value{
A data.table with simulated data, or a list with simulated data
  (data.dt), all separate effects (effects.dt), and summed effects
  (summed.effects.dt)
}
\description{
Creates a data.table with participants and simulates data for them.
}
