% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{generate.effects.dt}
\alias{generate.effects.dt}
\title{Creates a data.table of all effects that will be applied.}
\usage{
generate.effects.dt(
  data.dt,
  sim.parameters.dt,
  confound.interventions.parameters.dt
)
}
\arguments{
\item{data.dt}{data.table with minimally participant ID and cluster.}

\item{sim.parameters.dt}{The data.table generated by
generate.sim.parameters.dt}

\item{confound.interventions.parameters.dt}{The data.table generated by
generate.confound.interventions.parameters.dt}
}
\value{
A data.table including participant IDs, effect names, effect
  sizes, and the order in which they will be applied.
}
\description{
Creates a data.table of all effects that will be applied.
}
