% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{generate.flat.effect}
\alias{generate.flat.effect}
\title{A difference that is flat, special case of generate.step.effect where
step.effect is right at zero.}
\usage{
generate.flat.effect(time.vector, diff.amplitude = 1)
}
\arguments{
\item{time.vector}{A numeric vector of time values}

\item{diff.amplitude}{The numeric amplitude of the difference (Default: 1)}
}
\value{
A numeric vector containing the effect at each provided time point.
}
\description{
A difference that is flat, special case of generate.step.effect where
step.effect is right at zero.
}
