% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{generate.individual.noise}
\alias{generate.individual.noise}
\title{Generates random normal noise. Doesn't really need a time vector, other than
to know how many data points there are.}
\usage{
generate.individual.noise(time.vector, noise.sd = 0.5)
}
\arguments{
\item{time.vector}{A numeric vector of time values}

\item{noise.sd}{Standard deviation of the noise (Default: 0.5)}
}
\value{
A numeric vector containing the effect at each provided time point.
}
\description{
Generates random normal noise. Doesn't really need a time vector, other than
to know how many data points there are.
}
