% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{generate.linear.effect}
\alias{generate.linear.effect}
\title{Generate a difference that varies linearly over time. Transition times can be
equal, in which case there will be no transition (see generate.step.effect).
If both are zero, it is effectively a flat difference applied at every time
point (see generate.flat.effect).}
\usage{
generate.linear.effect(
  time.vector,
  diff.amplitude = 1,
  transition.start.time = 0,
  transition.end.time = 0
)
}
\arguments{
\item{time.vector}{A numeric vector of time values}

\item{diff.amplitude}{The numeric amplitude of the difference (Default: 1)}

\item{transition.start.time}{When does the effect start, will not be applied
if NA. (Default: 0)}

\item{transition.end.time}{When does the effect end, will not be applied
if NA. (Default: 0)}
}
\value{
A numeric vector containing the effect at each provided time point.
}
\description{
Generate a difference that varies linearly over time. Transition times can be
equal, in which case there will be no transition (see generate.step.effect).
If both are zero, it is effectively a flat difference applied at every time
point (see generate.flat.effect).
}
