% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{generate.sequence.dt}
\alias{generate.sequence.dt}
\title{Generates a data.table that describes the sequence of interventions in
clusters.}
\usage{
generate.sequence.dt(
  n.sequence,
  duration.of.period,
  transition.period = 0,
  intervention.sequence = F,
  control.sequence = F
)
}
\arguments{
\item{n.sequence}{Integer number of sequences.}

\item{duration.of.period}{Integer length of period.}

\item{transition.period}{Integer transition period. This is the number of
days before intervention that are considered transitional (Default: 0, no
transition)}

\item{intervention.sequence}{Boolean add a sequence that is entirely
intervention (Default: F)}

\item{control.sequence}{Boolean add a sequence that is entirely
control (Default: F)}
}
\value{
A data.table with sequence number, transition period start, and
  intervention start.
}
\description{
Generates a data.table that describes the sequence of interventions in
clusters.
}
