% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{generate.stat.dt}
\alias{generate.stat.dt}
\title{Construct a statistic distribution generated by permuting sites to different
clusters.}
\usage{
generate.stat.dt(
  data.dt,
  cluster.dt,
  sequence.dt,
  comparison.within = c("cluster", "period")[1],
  perm.dt = NULL,
  max.r = 1000,
  sort.input = T,
  exclude.transition = T,
  stat.func = c(test.wilcox.dt, test.f.effect.dt)[[1]],
  progress.bar = T
)
}
\arguments{
\item{data.dt}{data.table with columns participant, cluster, time, and
outcome. Outcome should be continuous.}

\item{cluster.dt}{data.table with the correspondence between cluster and
sequence, with columns cluster and sequence.}

\item{sequence.dt}{data.table with information about the sequences, with
columns sequence, transition.time, and intervention.time.}

\item{comparison.within}{Will comparisons be within cluster (i.e. between
period), or within period (i.e. within cluster).}

\item{perm.dt}{Precalculated table of how to permute clusters to sequences,
with first column being clusters, and then one column for each permutation
after that.}

\item{max.r}{How many permutations?}

\item{sort.input}{Will sort the input by outcome (by reference), which
slightly speeds ranking}

\item{exclude.transition}{boolean, should the result exclude data points from
the transition period? (Default = T)}

\item{stat.func}{What statistic will be used, there is a wilcox and f.effect
as I'm writing this. (Default: Wilcox)}

\item{progress.bar}{Display a progress bar. A little bit of overhead.
Completion times will be echoed regardless.}

\item{...}{Passed on to the test}
}
\value{
A data.table with the statistic value at each permutation (with zero
  as the unpermuted comparison).
}
\description{
Statistic is Wilcoxon-Mann-Whitney at the moment.
}
