% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.helpers.R
\name{generate.step.effect}
\alias{generate.step.effect}
\title{A difference that is a step, special case of generate.linear.effect where
transition.start and transition.end are equal.}
\usage{
generate.step.effect(time.vector, diff.amplitude = 1, step.time = 0)
}
\arguments{
\item{time.vector}{A numeric vector of time values}

\item{diff.amplitude}{The numeric amplitude of the difference (Default: 1)}

\item{step.time}{When does the step occur (Default: 0)}
}
\value{
A numeric vector containing the effect at each provided time point.
}
\description{
A difference that is a step, special case of generate.linear.effect where
transition.start and transition.end are equal.
}
