% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{icc.simulation.2D.scatter.plot}
\alias{icc.simulation.2D.scatter.plot}
\title{Creates a scatter plot of ICC vs one variable, ICC on y-axis.}
\usage{
icc.simulation.2D.scatter.plot(icc.dt, var.to.plot, plot.mean = T)
}
\arguments{
\item{icc.dt}{A data.table with the parameters of the various simulations and
ICC.}

\item{var.to.plot}{String name of the variable that will be on the x-axis.}

\item{plot.mean}{If true, plot the mean value at each point of var.to.plot,
otherwise plot all values individually.}
}
\value{
A scatter ggplot
}
\description{
Creates a scatter plot of ICC vs one variable, ICC on y-axis.
}
