% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{icc.simulation.3D.scatter.plot}
\alias{icc.simulation.3D.scatter.plot}
\title{Creates a scatter plot of ICC vs two other variables, ICC on y-axis.}
\usage{
icc.simulation.3D.scatter.plot(icc.dt, vars.to.plot, plot.mean = T)
}
\arguments{
\item{icc.dt}{A data.table with the parameters of the various simulations and
ICC.}

\item{vars.to.plot}{Vector of two string names of the variable that will be on
the x- and y-axes.}

\item{plot.mean}{If true, plot the mean value at each point of var.to.plot,
otherwise plot all values individually.}
}
\value{
A plotly 3D scatter plot
}
\description{
Creates a scatter plot of ICC vs two other variables, ICC on y-axis.
}
