% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpers.R
\name{icc.simulation.scatter.plot}
\alias{icc.simulation.scatter.plot}
\title{Creates a scatter plot of ICC vs one or two variables, ICC on y-axis. Chooses
automatically between a 3D plotly and 2D ggplot. If only one or two variables
in the parameters has multiple levels across scenarios, these will be
automatically detected.}
\usage{
icc.simulation.scatter.plot(
  icc.dt,
  vars.to.plot = NULL,
  plot.mean = T,
  cols.to.ignore = c("iteration", "icc")
)
}
\arguments{
\item{icc.dt}{A data.table with the parameters of the various simulations and
ICC.}

\item{vars.to.plot}{String name of the variable that will be on the x-axis.
Will autodetect if NULL, but that will fail if there are more than three
potential variables to plot ICC against. (Default: NULL)}

\item{plot.mean}{If true, plot the mean value at each point of var.to.plot,
otherwise plot all values individually.}

\item{cols.to.ignore}{Columns that will be ignored by detection of
vars.to.plot (Default: c('iteration', 'icc'))}
}
\value{
A scatter ggplot
}
\description{
Creates a scatter plot of ICC vs one or two variables, ICC on y-axis. Chooses
automatically between a 3D plotly and 2D ggplot. If only one or two variables
in the parameters has multiple levels across scenarios, these will be
automatically detected.
}
