% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{perform.permutation.step}
\alias{perform.permutation.step}
\title{A function to perform a permutation step, so that the process can be deployed
in parallel.}
\usage{
perform.permutation.step(
  perm.ind,
  perm.dt,
  data.dt,
  cluster.dt,
  sequence.dt,
  hypothetical.data.dt,
  comparison.within = c("cluster", "period")[1],
  stat.func = c(test.wilcox.dt, test.f.effect.dt)[[1]],
  progress.bar = NULL
)
}
\arguments{
\item{perm.ind}{Index of permutation column in perm.dt that should be used.}

\item{perm.dt}{Precalculated table of how to permute clusters to sequences,
with first column being clusters, and then one column for each permutation
after that.}

\item{data.dt}{data.table with columns participant, cluster, time, and
outcome. Outcome should be continuous.}

\item{cluster.dt}{data.table with the correspondence between cluster and
sequence, with columns cluster and sequence.}

\item{sequence.dt}{data.table with information about the sequences, with
columns sequence, transition.time, and intervention.time.}

\item{hypothetical.data.dt}{data.table with hypothetical data point condition
assignments if each cluster were in each sequence. Three columns, sequence,
cluster (containing all combinations), and data.dt, containing the
hypothetical data.dt from only that cluster.}

\item{comparison.within}{Will comparisons be within cluster (i.e. between
period), or within period (i.e. within cluster).}

\item{stat.func}{What statistic will be used, there is a wilcox and f.effect
as I'm writing this. (Default: Wilcox)}

\item{progress.bar}{Should be a text progress bar if you want one.}
}
\value{
A data.table with the stat separated by period and permutation
  number.
}
\description{
A function to perform a permutation step, so that the process can be deployed
in parallel.
}
