% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{remove.rows.with.no.comparison}
\alias{remove.rows.with.no.comparison}
\title{Removes rows from periods/clusters that do not have participants in both the
control and intervention conditions. Needs the conditions attached, so will
attempt to do so if not already done.}
\usage{
remove.rows.with.no.comparison(
  data.dt,
  cluster.dt = NULL,
  sequence.dt = NULL,
  remove.type = c("cluster", "period")[1]
)
}
\arguments{
\item{data.dt}{data.table with columns participant, cluster, time, and
outcome. Outcome should be continuous.}

\item{cluster.dt}{data.table with the correspondence between cluster and
sequence, with columns cluster and sequence.}

\item{sequence.dt}{data.table with information about the sequences, with
columns sequence, transition.time, and intervention.time.}

\item{remove.type}{character indicating what to compare on.}
}
\value{
data.dt with rows removed that did not have a corresponding control
  or intervention.
}
\description{
Removes rows from periods/clusters that do not have participants in both the
control and intervention conditions. Needs the conditions attached, so will
attempt to do so if not already done.
}
