% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{test.f.effect.dt}
\alias{test.f.effect.dt}
\title{Do a difference in means test on a data.table, as outlined in Thompson,
Davey, et al 2018}
\usage{
test.f.effect.dt(data.dt)
}
\arguments{
\item{data.dt}{data.table with two-level grouping factor 'condition' and
outcome 'outcome'}
}
\value{
A data.table containing mean difference by condition for each period
  and a variance weighting for calculating the mean, along with means,
  variances, and numbers of periods for each level of condition.
}
\description{
Do a difference in means test on a data.table, as outlined in Thompson,
Davey, et al 2018
}
