% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.helpers.R
\name{test.wilcox.dt}
\alias{test.wilcox.dt}
\title{Do a wilcox test on a data.table. More efficient that coin for large numbers,
sort the input by outcome to make it faster.}
\usage{
test.wilcox.dt(data.dt, two.sided = T, tie.correction = NULL, sort.input = F)
}
\arguments{
\item{data.dt}{data.table with two-level grouping factor 'condition' and
outcome 'outcome''}

\item{two.sided}{Do a two sided test?}

\item{tie.correction}{Apply tie correction (slightly slower, but sometimes
necessary). If NULL, tie correction will be applied if there are.}

\item{sort.input}{Sorts the input by reference, this enables faster ranking
(i.e. speeds permutation test)}
}
\value{
A data.table containing statistics, including z score and theoretical
  p value (e.g. z.dt$z).
}
\description{
Do a wilcox test on a data.table. More efficient that coin for large numbers,
sort the input by outcome to make it faster.
}
