% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.helpers.R
\name{update.progress.bar}
\alias{update.progress.bar}
\title{Update a progress bar.}
\usage{
\method{update}{progress.bar}(progress.bar, index, max.r, max.n.updates = 50)
}
\arguments{
\item{progress.bar}{A text progress bar item}

\item{index}{Numeric index of the permutation you are in (out of max.r)}

\item{max.r}{Numeric how many you are expecting total}

\item{max.n.updates}{Maximum number of updates, there is a bit of overhead,
so don't want it updating every iteration. (Default: 50)}
}
\description{
Update a progress bar.
}
